+++ src/AbilityProviders/WardenAbilityProvider.php
@@ -21,6 +21,7 @@ use function resolve;
 
 /**
  * @psalm-immutable
+ * @author Brian Faust <brian@cline.sh>
  */
 final readonly class WardenAbilityProvider implements AbilityProviderInterface
 {